+++ layers/Dropout.py
@@ -22,7 +22,7 @@ class Dropout(Layer):
         self._drop_prob = 1
         # remember the star in front if inp.shape is to unpack the tuple
         if self.mode == Mode.TRAIN:
-            self._drop_prob = ((np.random.rand(*self._inp.shape) < self.probability) / self.probability)
+            self._drop_prob = ((np.float32(np.random.rand(*self._inp.shape)) < self.probability) / self.probability)
         return self._inp * self._drop_prob
 
     def backward_pass(self, upstream_grad):
